+++ schemas/rate_request.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel, Field, validator
 from typing import Optional, List
-from datetime import datetime
 
 from core.constants.ups import UPSConstants
 from core.logging import get_logger
@@ -44,7 +43,7 @@ class Package(BaseModel):
     return max(1.0, float(v))
 
 
-class UPSRateRequest(BaseModel):
+class RateRequest(BaseModel):
   sender_country: str = Field(..., description="Sender country code")
   country: str = Field(..., description="Destination country code")
   weight: float = Field(..., gt=0, description="Package weight")
@@ -154,22 +153,3 @@ class UPSRateRequest(BaseModel):
 
     if errors:
       raise ValueError("; ".join(errors))
-
-
-class RateQuote(BaseModel):
-  service_code: str
-  service_name: str
-  total_charge: float
-  currency_code: str
-  delivery_days: Optional[int] = None
-  guaranteed: Optional[bool] = None
-  context: Optional[str] = None
-  method_id: Optional[int] = None
-
-
-class RateResponse(BaseModel):
-  quotes: List[RateQuote]
-  request_id: str
-  timestamp: datetime
-  processing_time_ms: float
-  context: str
